move some things around, more docs
