+++ pennylane/ops/qubit/__init__.py
@@ -13,13 +13,25 @@
 # limitations under the License.
 
 """
-This submodule contains all qubit gates.
+This module contains the discrete-variable quantum operations.
+
+The operations are divided into seven different files:
+
+* ``non_parameteric_ops.py``: All operations with no parameters
+* ``parameteric_ops.py``: Core parameteric operations that don't fall into
+  any of the other categories
+* ``qchem_ops.py``: Operations for quantum chemistry applications
+* ``matrix_ops.py``: Generalized operations that accept a matrix parameter,
+  either unitary or hermitian depending
+* ``arithmetic_ops.py``: Operations that perform arithmetic on the states
+* ``state_preparation.py``: Operations that initialize the state
+
 """
 
 from .non_parametric_ops import *
 from .parametric_ops import *
 from .qchem_ops import *
-from .arbitrary_ops import *
+from .matrix_ops import *
 from .arthimetic_ops import *
 from .state_preparation import *
 from .other import *
